requirements.txt
absl-py==2.1.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
Bottleneck @ file:///croot/bottleneck_1707864210935/work
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
cachetools==5.3.3
certifi @ file:///croot/certifi_1707229174982/work/certifi
cffi @ file:///croot/cffi_1700254295673/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
comm==0.2.2
contourpy @ file:///croot/contourpy_1700583582875/work
cryptography @ file:///croot/cryptography_1702070282333/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
d2l==1.0.3
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.19.1
filelock @ file:///croot/filelock_1700591183607/work
fonttools @ file:///croot/fonttools_1713551344105/work
fqdn==1.5.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
google-auth==2.29.0
google-auth-oauthlib==0.4.6
grpcio==1.62.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.0
idna @ file:///croot/idna_1666125576474/work
ipykernel==6.29.4
ipython==8.23.0
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2 @ file:///croot/jinja2_1666908132255/work
joblib @ file:///croot/joblib_1713976767370/work
json5==0.9.25
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.26.0
jupyterlab_widgets==3.0.10
kaggle==1.6.12
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
Markdown==3.6
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==3.0.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx @ file:///croot/networkx_1690561992265/work
notebook==7.1.3
notebook_shim==0.2.4
numexpr @ file:///croot/numexpr_1696515281613/work
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.9.0.80
overrides==7.7.0
packaging @ file:///croot/packaging_1710807400464/work
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
Pillow @ file:///croot/pillow_1696580024257/work
platformdirs==4.2.0
ply==3.11
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.17.2
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PyQt5==5.15.10
PyQt5-sip @ file:///croot/pyqt-split_1698769088074/work/pyqt_sip
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-json-logger==2.0.7
python-slugify==8.0.4
pytz @ file:///croot/pytz_1713974312559/work
PyYAML @ file:///croot/pyyaml_1698096049011/work
pyzmq==26.0.0
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.34.0
requests @ file:///croot/requests_1690400202158/work
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.3
scikit-learn @ file:///croot/scikit-learn_1694788527225/work
scipy==1.10.1
Send2Trash==1.8.3
sentry-sdk==2.1.1
setproctitle==1.3.3
sip @ file:///croot/sip_1698675935381/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
sympy @ file:///croot/sympy_1701397643339/work
tensorboard==2.12.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
terminado==0.18.1
text-unidecode==1.3
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
timm==0.9.16
tinycss2==1.2.1
tomli @ file:///opt/conda/conda-bld/tomli_1657175507142/work
torch==2.0.0
torchaudio==2.0.0
torchvision==0.15.0
tornado @ file:///croot/tornado_1696936946304/work
tqdm==4.66.2
traitlets==5.14.2
triton==2.0.0
types-python-dateutil==2.9.0.20240316
typing_extensions @ file:///croot/typing_extensions_1690297465030/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
unicodedata2 @ file:///croot/unicodedata2_1713212950228/work
uri-template==1.3.0
urllib3 @ file:///croot/urllib3_1698257533958/work
wandb==0.17.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.2
widgetsnbextension==4.0.10
